// src/core/html-report/types.ts
export interface ReportTemplate {
  /** Template name */
  name: string;
  
  /** Template content */
  content: string;
  
  /** Template type */
  type: 'main' | 'section' | 'chart' | 'partial';
  
  /** Template variables */
  variables: string[];
  
  /** Template description */
  description?: string;
  
  /** Template author */
  author?: string;
  
  /** Template version */
  version?: string;
}